additional refactor task
