+++ src/rtl_sdr_fm.cpp
@@ -34,6 +34,7 @@
 #include <unistd.h>
 #include "dsd.h"
 #include "dsp/demod_pipeline.h"
+#include "dsp/demod_state.h"
 #include "dsp/fll.h"
 #include "dsp/resampler.h"
 #include "dsp/ted.h"
@@ -280,122 +281,6 @@ struct dongle_state {
     int direct_sampling;
     std::atomic<int> mute;
     struct demod_state* demod_target;
-};
-
-struct demod_state {
-    int exit_flag;
-    pthread_t thread;
-    int16_t* lowpassed;
-    /* Scratch buffer for demod thread to read blocks from the input ring */
-    /* Not a ring; callback writes directly into the global input ring. */
-    alignas(DSD_FME_ALIGN) int16_t input_cb_buf[MAXIMUM_BUF_LENGTH];
-    int lp_len;
-    int16_t lp_i_hist[10][6];
-    int16_t lp_q_hist[10][6];
-    alignas(DSD_FME_ALIGN) int16_t result[MAXIMUM_BUF_LENGTH];
-    int16_t droop_i_hist[9];
-    int16_t droop_q_hist[9];
-    int result_len;
-    int rate_in;
-    int rate_out;
-    int rate_out2;
-    int now_r, now_j;
-    int pre_r, pre_j;
-    int prev_index;
-    int downsample; /* min 1, max 256 */
-    int post_downsample;
-    int output_scale;
-    int squelch_level, conseq_squelch, squelch_hits, terminate_on_squelch;
-    /* Incremental, decimated RMS squelch estimator (power-domain, sqrt-free) */
-    int64_t squelch_running_power;
-    int squelch_decim_stride;
-    int squelch_decim_phase;
-    int squelch_window;
-    int downsample_passes;
-    int comp_fir_size;
-    int custom_atan;
-    int deemph, deemph_a;
-    int deemph_avg;
-    /* Optional post-demod audio low-pass filter (one-pole) */
-    int audio_lpf_enable;
-    int audio_lpf_alpha; /* Q15 alpha for one-pole LPF */
-    int audio_lpf_state; /* state/output y[n-1] in Q0 */
-    int now_lpr;
-    int prev_lpr_index;
-    int dc_block, dc_avg;
-    /* Half-band decimator state */
-    int16_t hb_workbuf[MAXIMUM_BUF_LENGTH];
-    int16_t hb_hist_i[10][HB_TAPS - 1];
-    int16_t hb_hist_q[10][HB_TAPS - 1];
-    alignas(DSD_FME_ALIGN) int16_t hb_i_buf[MAXIMUM_BUF_LENGTH / 2];
-    alignas(DSD_FME_ALIGN) int16_t hb_q_buf[MAXIMUM_BUF_LENGTH / 2];
-    alignas(DSD_FME_ALIGN) int16_t hb_i_out[MAXIMUM_BUF_LENGTH / 2];
-    alignas(DSD_FME_ALIGN) int16_t hb_q_out[MAXIMUM_BUF_LENGTH / 2];
-    /* Preallocated buffer for linear upsampler (bandwidth_multiplier) */
-    alignas(DSD_FME_ALIGN) int16_t upsample_buf[MAXIMUM_BUF_LENGTH * MAX_BANDWIDTH_MULTIPLIER];
-    /* Polyphase rational resampler (L/M) state and output buffer */
-    int resamp_enabled;
-    int resamp_target_hz;      /* desired output sample rate */
-    int resamp_L;              /* upsample factor */
-    int resamp_M;              /* downsample factor */
-    int resamp_phase;          /* 0..L-1 accumulator */
-    int resamp_taps_len;       /* prototype taps length (padded to K*L) */
-    int resamp_taps_per_phase; /* K = ceil(taps_len/L) */
-    int16_t* resamp_taps;      /* Q15 taps, length = K*L */
-    int16_t* resamp_hist;      /* circular history, length = K */
-    int resamp_hist_head;      /* head index into circular history [0..K-1] */
-    /* Output buffer for resampler (worst-case 4x expansion) */
-    alignas(DSD_FME_ALIGN) int16_t resamp_outbuf[MAXIMUM_BUF_LENGTH * 4];
-    /* Residual CFO loop (FLL) state */
-    int fll_enabled;
-    int fll_alpha_q15;    /* proportional gain (Q15) */
-    int fll_beta_q15;     /* integral gain (Q15) */
-    int fll_freq_q15;     /* NCO frequency increment (Q15 radians/sample scaled) */
-    int fll_phase_q15;    /* NCO phase accumulator (wrap at 2*pi -> 1<<15 scale) */
-    int fll_deadband_q14; /* ignore small phase errors |err| <= deadband (Q14) */
-    int fll_slew_max_q15; /* max |delta freq| per update (Q15) */
-    int fll_prev_r;
-    int fll_prev_j;
-    /* Timing error detector (Gardner) fractional-delay state */
-    int ted_enabled;
-    int ted_force;    /* allow forcing TED even for FM/C4FM paths */
-    int ted_gain_q20; /* small gain (Q20) for stability */
-    int ted_sps;      /* nominal samples per symbol (e.g., 10 for 4800 sym/s at 48k) */
-    int ted_mu_q20;   /* fractional phase [0,1) in Q20 */
-    /* Work buffer for timing-adjusted I/Q */
-    alignas(DSD_FME_ALIGN) int16_t timing_buf[MAXIMUM_BUF_LENGTH];
-    /* FLL and TED module states */
-    fll_state_t fll_state;
-    ted_state_t ted_state;
-    /* Minimal 2-thread worker pool for intra-block parallelism */
-    int mt_enabled;
-    int mt_ready;
-    pthread_t mt_threads[2];
-    pthread_mutex_t mt_lock;
-    pthread_cond_t mt_cv;
-    pthread_cond_t mt_done_cv;
-    int mt_should_exit;
-    int mt_epoch;
-    int mt_completed_in_epoch;
-    int mt_posted_count;
-
-    struct {
-        void (*run)(void*);
-        void* arg;
-    } mt_tasks[2];
-
-    int mt_worker_id[2];
-
-    struct {
-        struct demod_state* s;
-        int id;
-    } mt_args[2];
-
-    int (*discriminator)(int, int, int, int);
-    void (*mode_demod)(struct demod_state*);
-    pthread_cond_t ready;
-    pthread_mutex_t ready_m;
-    struct output_state* output_target;
 };
 
 struct demod_mt_worker_arg {
@@ -663,101 +548,6 @@ polar_disc_lut(int ar, int aj, int br, int bj) {
     }
 
     return 0;
-}
-
-/**
- * Mix lowpassed I/Q by NCO e^{j*phi}, update phase by `fll_freq_q15` per sample.
- * Phase and frequency are Q15 where a full turn (2*pi) maps to 1<<15.
- *
- * @param d Demodulator state (reads/writes `lowpassed`, updates `fll_phase_q15`).
- */
-static inline void
-fll_mix_and_update(struct demod_state* d) {
-    if (!d->fll_enabled) {
-        return;
-    }
-
-    /* Sync from demod_state to module state */
-    d->fll_state.freq_q15 = d->fll_freq_q15;
-    d->fll_state.phase_q15 = d->fll_phase_q15;
-    d->fll_state.prev_r = d->fll_prev_r;
-    d->fll_state.prev_j = d->fll_prev_j;
-
-    fll_config_t cfg = {.enabled = d->fll_enabled,
-                        .alpha_q15 = d->fll_alpha_q15,
-                        .beta_q15 = d->fll_beta_q15,
-                        .deadband_q14 = d->fll_deadband_q14,
-                        .slew_max_q15 = d->fll_slew_max_q15,
-                        .use_lut = fll_lut_enabled};
-
-    fll_mix_and_update(&cfg, &d->fll_state, d->lowpassed, d->lp_len);
-
-    /* Sync back to demod_state */
-    d->fll_freq_q15 = d->fll_state.freq_q15;
-    d->fll_phase_q15 = d->fll_state.phase_q15;
-    d->fll_prev_r = d->fll_state.prev_r;
-    d->fll_prev_j = d->fll_state.prev_j;
-}
-
-/**
- * Estimate frequency error using a simple phase-difference discriminator and
- * update the FLL control in Q15. The proportional term is applied directly
- * and the integral action is realized by accumulating into `fll_freq_q15`.
- *
- * @param d Demodulator state (updates `fll_freq_q15` and `fll_phase_q15`).
- */
-static inline void
-fll_update_error(struct demod_state* d) {
-    if (!d->fll_enabled) {
-        return;
-    }
-
-    /* Sync from demod_state to module state */
-    d->fll_state.freq_q15 = d->fll_freq_q15;
-    d->fll_state.phase_q15 = d->fll_phase_q15;
-    d->fll_state.prev_r = d->fll_prev_r;
-    d->fll_state.prev_j = d->fll_prev_j;
-
-    fll_config_t cfg = {.enabled = d->fll_enabled,
-                        .alpha_q15 = d->fll_alpha_q15,
-                        .beta_q15 = d->fll_beta_q15,
-                        .deadband_q14 = d->fll_deadband_q14,
-                        .slew_max_q15 = d->fll_slew_max_q15,
-                        .use_lut = fll_lut_enabled};
-
-    fll_update_error(&cfg, &d->fll_state, d->lowpassed, d->lp_len);
-
-    /* Sync back to demod_state */
-    d->fll_freq_q15 = d->fll_state.freq_q15;
-    d->fll_phase_q15 = d->fll_state.phase_q15;
-    d->fll_prev_r = d->fll_state.prev_r;
-    d->fll_prev_j = d->fll_state.prev_j;
-}
-
-/**
- * Lightweight Gardner timing correction.
- * Uses linear interpolation between adjacent complex samples around the
- * nominal samples-per-symbol to reduce timing error; intended for digital
- * modes when enabled.
- *
- * @param d Demodulator state (may adjust `result` in-place).
- */
-static inline void
-gardner_timing_adjust(struct demod_state* d) {
-    if (!d->ted_enabled || d->ted_sps <= 1) {
-        return;
-    }
-
-    /* Sync from demod_state to module state */
-    d->ted_state.mu_q20 = d->ted_mu_q20;
-
-    ted_config_t cfg = {
-        .enabled = d->ted_enabled, .force = d->ted_force, .gain_q20 = d->ted_gain_q20, .sps = d->ted_sps};
-
-    gardner_timing_adjust(&cfg, &d->ted_state, d->lowpassed, &d->lp_len, d->timing_buf);
-
-    /* Sync back to demod_state */
-    d->ted_mu_q20 = d->ted_state.mu_q20;
 }
 
 /**
